refactor(control): split capture-with-actions into commander capture group + lib modules
